refactor(product): 기존 도메인 리팩토링 + 관련된 사이드 이펙트 수정 및 테스트 정상 작동

+++ src/main/java/kitchenpos/core/shared/data/jpa/MoneyConverter.java
@@ -4,15 +4,17 @@ import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 import kitchenpos.core.shared.value.Money;
 
+import java.math.BigDecimal;
+
 @Converter(autoApply = true)
-public class MoneyConverter implements AttributeConverter<Money, Long> {
+public class MoneyConverter implements AttributeConverter<Money, BigDecimal> {
     @Override
-    public Long convertToDatabaseColumn(Money money) {
-        return money.getAmount().longValue();
+    public BigDecimal convertToDatabaseColumn(Money money) {
+        return money.getAmount();
     }
 
     @Override
-    public Money convertToEntityAttribute(Long amount) {
+    public Money convertToEntityAttribute(BigDecimal amount) {
         return Money.wons(amount);
     }
 }
